Manual file select now possible - tests not in place yet
